Rename CentCom to Api
- In anticipation for exposing webpages too.
