+++ Api/Controllers/UserController.cs
@@ -2,16 +2,16 @@ using System;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
-using CentCom.Dtos;
-using CentCom.Helpers;
-using CentCom.Interfaces;
-using CentCom.Models;
+using Api.Dtos;
+using Api.Helpers;
+using Api.Interfaces;
+using Api.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 
-namespace CentCom.Controllers
+namespace Api.Controllers
 {
     [Authorize]
     [Route("api/[controller]")]
