+++ Api/Interfaces/ICharacterService.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
-using CentCom.Models;
+using Api.Models;
 
-namespace CentCom.Interfaces
+namespace Api.Interfaces
 {
     public interface ICharacterService
     {
