+++ Api/Program.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 
-namespace CentCom
+namespace Api
 {
     public class Program
     {
