+++ Api/Services/CharacterService.cs
@@ -1,10 +1,10 @@
 using System.Collections.Generic;
 using System.Linq;
-using CentCom.Helpers;
-using CentCom.Interfaces;
-using CentCom.Models;
+using Api.Helpers;
+using Api.Interfaces;
+using Api.Models;
 
-namespace CentCom.Services
+namespace Api.Services
 {
     public class CharacterService : ICharacterService
     {
