+++ Api/appsettings.Development.json
@@ -5,8 +5,5 @@
             "System": "Information",
             "Microsoft": "Information"
         }
-    },
-    "ConnectionStrings": {
-        "DefaultConnection": "Server=localhost;Port=3306;Database=DATABASE_NAME;User Id=DATABASE_USER;Password=PASSWORD"
     }
 }
